added automatic updating terms in term channels

+++ config/app_config.py
@@ -157,7 +157,7 @@ class Config:
     exams_page_size:int = get_attr(toml_dict, "exams", "page_size")
     exams_paginator_duration:int = get_attr(toml_dict, "exams", "paginator_duration")
     exams_term_channels:List[str] = get_attr(toml_dict, "exams", "term_channels")
-    exams_terms_update_interval:int = get_attr(toml_dict, "exams", "terms_update_interval")
+    exams_terms_update_interval:float = get_attr(toml_dict, "exams", "terms_update_interval")
 
 config = Config()
